+++ src/main/java/org/bytedeco/javacpp/tools/InfoMap.java
@@ -93,8 +93,8 @@ public class InfoMap extends HashMap<String,List<Info>> {
         .put(new Info("const char").valueTypes("byte").pointerTypes("@Cast(\"const char*\") BytePointer", "String"))
         .put(new Info("boost::shared_ptr", "std::shared_ptr").annotations("@SharedPtr"))
         .put(new Info("boost::movelib::unique_ptr", "std::unique_ptr").annotations("@UniquePtr"))
-        .put(new Info("std::string").annotations("@StdString").valueTypes("BytePointer", "String"))
-        .put(new Info("std::wstring").annotations("@StdWString").valueTypes("CharPointer", "IntPointer"))
+        .put(new Info("std::string").annotations("@StdString").valueTypes("BytePointer", "String").pointerTypes("BytePointer"))
+        .put(new Info("std::wstring").annotations("@StdWString").valueTypes("CharPointer", "IntPointer").pointerTypes("CharPointer", "IntPointer"))
         .put(new Info("std::vector").annotations("@StdVector"))
 
         .put(new Info("abstract").javaNames("_abstract"))